rename: HasSnowGlobe ➡  HasSnowball
