+++ front/src/router/index.tsx
@@ -1,2 +1,2 @@
 export { default as IsLogin } from './IsLogin';
-export { default as HasSnowball } from './HasSnowGlobe';
+export { default as HasSnowball } from './HasSnowball';
